add ApolloClient.Builder.retryOnErrorInterceptor
